Create mongotest.rb for the email_handler
Committing a helper file that pulls a group out of your mongodb and makes an array of that group's members. The file contains a bunch of tests and example methods that will be useful later.
This array will be used by email_handler to send encrypted mail to each group member with their key.